beacon e2e integration test
