+++ Cargo.toml
@@ -12,6 +12,8 @@ members = [
     "iso-aggregate",
 ]
 
+default-members = ["prover"]
+
 # Always optimize; building and running the guest takes much longer without optimization.
 [profile.dev]
 opt-level = 3
