+++ prover/Cargo.toml
@@ -14,6 +14,8 @@ serde_json = "1"
 committee-iso = { path = "../iso-committee" }
 step-iso = { path = "../iso-step" }
 aggregate-iso = { path = "../iso-aggregate" }
+preprocessor = { path = "../preprocessor" }
+eth-types = { path = "../eth-types" }
 aligned-sdk = { git = "https://github.com/yetanotherco/aligned_layer", tag = "v0.9.2" }
 hex = "0.4.0"
 rpassword = "7.3.1"
