--- prover/src/integrations/aligned/constants.rs
@@ -1,2 +1,0 @@
-pub const BATCHER_URL: &str = "wss://batcher.alignedlayer.com";
-pub const ETH_RPC_URL: &str = "https://ethereum-holesky-rpc.publicnode.com";
