--- prover/src/integrations/mod.rs
@@ -1,1 +1,0 @@
-pub mod aligned;
