+++ prover/src/lib.rs
@@ -1,8 +1,13 @@
+use aggregate_iso::types::RecursiveInputs;
 use committee_iso::types::CommitteeUpdateArgs;
-use sp1_sdk::{include_elf, ProverClient, SP1ProofWithPublicValues, SP1Stdin, SP1VerifyingKey};
+use sp1_sdk::{
+    include_elf, ProverClient, SP1Proof, SP1ProofWithPublicValues, SP1Stdin, SP1VerifyingKey,
+};
 use step_iso::types::{SyncStepArgs, SyncStepCircuitInput};
-
-pub mod integrations;
+pub enum ProofCompressionBool {
+    Compressed,
+    Uncompressed,
+}
 
 pub enum ProverOps {
     Default,
@@ -13,6 +18,7 @@ pub enum ProverOps {
 pub fn generate_committee_update_proof_sp1(
     ops: &ProverOps,
     committee_update: CommitteeUpdateArgs,
+    compressed: &ProofCompressionBool,
 ) -> (SP1ProofWithPublicValues, SP1VerifyingKey) {
     use std::time::Instant;
     let start_time = Instant::now();
@@ -24,10 +30,17 @@ pub fn generate_committee_update_proof_sp1(
         ProverOps::Default => {
             const COMMITTEE_ELF: &[u8] = include_elf!("sp1-committee");
             let (pk, vk) = client.setup(COMMITTEE_ELF);
-            let proof = client
-                .prove(&pk, stdin)
-                .run()
-                .expect("failed to generate proof");
+            let proof = match compressed {
+                ProofCompressionBool::Compressed => client
+                    .prove(&pk, stdin)
+                    .compressed()
+                    .run()
+                    .expect("failed to generate proof"),
+                ProofCompressionBool::Uncompressed => client
+                    .prove(&pk, stdin)
+                    .run()
+                    .expect("failed to generate proof"),
+            };
             (proof, pk, vk)
         }
         ProverOps::Groth16 => {
@@ -61,6 +74,7 @@ pub fn generate_step_proof_sp1(
     ops: &ProverOps,
     commitment: [u8; 32],
     sync_step_args: SyncStepArgs,
+    compressed: &ProofCompressionBool,
 ) -> (SP1ProofWithPublicValues, SP1VerifyingKey) {
     use std::time::Instant;
     sp1_sdk::utils::setup_logger();
@@ -77,10 +91,17 @@ pub fn generate_step_proof_sp1(
         ProverOps::Default => {
             const STEP_ELF: &[u8] = include_elf!("sp1-step");
             let (pk, vk) = client.setup(STEP_ELF);
-            let proof = client
-                .prove(&pk, stdin)
-                .run()
-                .expect("failed to generate proof");
+            let proof = match compressed {
+                ProofCompressionBool::Compressed => client
+                    .prove(&pk, stdin)
+                    .compressed()
+                    .run()
+                    .expect("failed to generate proof"),
+                ProofCompressionBool::Uncompressed => client
+                    .prove(&pk, stdin)
+                    .run()
+                    .expect("failed to generate proof"),
+            };
             (proof, pk, vk)
         }
         ProverOps::Groth16 => {
@@ -105,6 +126,62 @@ pub fn generate_step_proof_sp1(
         }
     };
     println!("Successfully generated proof!");
+    let duration = start_time.elapsed();
+    println!("Elapsed time: {:?}", duration);
+    (proof, vk)
+}
+
+pub fn generate_aggregate_proof_sp1(
+    ops: &ProverOps,
+    inputs: Vec<RecursiveInputs>,
+    proofs: Vec<SP1ProofWithPublicValues>,
+    vks: Vec<SP1VerifyingKey>,
+) -> (SP1ProofWithPublicValues, SP1VerifyingKey) {
+    use std::time::Instant;
+    sp1_sdk::utils::setup_logger();
+    let start_time = Instant::now();
+    let client = ProverClient::new();
+    let mut stdin = SP1Stdin::new();
+    stdin.write_vec(borsh::to_vec(&inputs).expect("Failed to serialize"));
+    // write first proof - committee
+    let SP1Proof::Compressed(proof) = proofs.first().unwrap().proof.clone() else {
+        panic!("Uncompressed proof unsupported: Committee!")
+    };
+    stdin.write_proof(*proof, vks.first().unwrap().vk.clone());
+    // write second proof - committee
+    let SP1Proof::Compressed(proof) = proofs.get(1).unwrap().proof.clone() else {
+        panic!("Uncompressed proof unsupported: Step!")
+    };
+    stdin.write_proof(*proof, vks.get(1).unwrap().vk.clone());
+
+    let (proof, _, vk) = match ops {
+        ProverOps::Default => {
+            panic!("Recursive Step Proof for Default Prover mode is not supported!")
+        }
+        ProverOps::Groth16 => {
+            const RECURSIVE_ELF: &[u8] = include_elf!("sp1-aggregate");
+            let (pk, vk) = client.setup(RECURSIVE_ELF);
+            let proof = client
+                .prove(&pk, stdin)
+                .groth16()
+                .run()
+                .expect("failed to generate proof");
+            (proof, pk, vk)
+        }
+        ProverOps::Plonk => {
+            const RECURSIVE_ELF: &[u8] = include_elf!("sp1-aggregate");
+            let (pk, vk) = client.setup(RECURSIVE_ELF);
+            let proof = client
+                .prove(&pk, stdin)
+                .plonk()
+                .run()
+                .expect("failed to generate proof");
+            (proof, pk, vk)
+        }
+    };
+    println!("Successfully generated proof!");
+    client.verify(&proof, &vk).expect("failed to verify proof");
+    println!("Successfully verified proof!");
     let duration = start_time.elapsed();
     println!("Elapsed time: {:?}", duration);
     (proof, vk)
